Add history model to track firefighter status changes
This PR reorganizes the modules in multiple directories:
* config - db and env stuff
* models - mongoose model definitions
* repo - repo modules to save, update and list our models
* web - express web server stuff

Then it adds a new model `History` that is associated with one Firefighter and contains a `before` and `after` field, to keep track of changes. Every time we update the status of a firefighter, the firefighter repo calls a method on the history repo to create a new history record.

With this we are able to build analytics.